Refactor engine to handle lack board info

+++ src/main.rs
@@ -131,7 +131,7 @@ impl App {
                                 match cell {
                                     Cell::Figure(figure) => {
                                         let moves =
-                                            self.game.possible_moves(rank as u32, file as u32);
+                                            self.game.possible_moves(rank, file);
                                         self.chosen_figure = if moves.is_some() {
                                             self.selected_cell = Some((rank, file));
                                             Some(figure.clone())
